fixed some compiler warnings and improved CppADCGPatternTest

+++ test/cppadcg/dae_index_reduction/dummy_derivative.cpp
@@ -57,7 +57,7 @@ TEST_F(CppADCGIndexReductionTest, DummyDerivPendulum2D) {
 
     ASSERT_TRUE(reducedFun != NULL);
 
-    ASSERT_EQ(3, dummyD.getDifferentiationIndex());
+    ASSERT_EQ(size_t(3), dummyD.getDifferentiationIndex());
 
     delete fun;
     delete reducedFun;
